Orders history (#8)
- darkened store section background color,
- implemented redirect from order item to store,
- implemented order group info popup (with browser redirects to recipe & support page),
- implemented proper order replies handling,
- implemented error page for orders,
- made use of system locale for order group displayed info

+++ api/utils/orderserialization.cpp
@@ -4,8 +4,8 @@
 
 void parseProductPrice(const QJsonObject &json, api::OrderProductPrice &data)
 {
-    data.baseAmount = json["baseAmount"].toString();
-    data.amount = json["amount"].toString();
+    data.baseAmount = json["baseAmount"].toString().toDouble();
+    data.amount = json["amount"].toString().toDouble();
     data.free = json["isFree"].toBool();
     data.discounted = json["isDiscounted"].toBool();
     data.symbol = json["symbol"].toString();
@@ -13,7 +13,7 @@ void parseProductPrice(const QJsonObject &json, api::OrderProductPrice &data)
 
 void parseOrderPrice(const QJsonObject &json, api::OrderPrice &data)
 {
-    data.amount = json["amount"].toString();
+    data.amount = json["amount"].toString().toDouble();
     data.symbol = json["symbol"].toString();
     data.code = json["code"].toString();
     data.zero = json["isZero"].toBool();
